refactor(actions): refactor document actions
