+++ packages/web-app/src/actions/Document/ResetApiMessages.js
@@ -1,0 +1,5 @@
+export const RESET_API_MESSAGES = 'RESET_API_MESSAGES';
+
+export const resetApiMessages = () => ({
+  type: RESET_API_MESSAGES
+});
