+++ packages/web-app/src/components/appli/Duplicates/DuplicatesList.jsx
@@ -14,7 +14,7 @@ import Table from '../../common/Table';
 import useMakeCustomHeaderCellRenders from './customHeaderCellRender';
 import useMakeCustomCellRenders from './customCellRender';
 import TableActions from './TableActions';
-import { RESET_API_MESSAGES } from '../../../actions/Document';
+import { resetApiMessages } from '../../../actions/Document/ResetApiMessages';
 import { resetEntranceState } from '../../../actions/Entrance/ResetEntrance';
 
 const Wrapper = styled.div`
@@ -39,7 +39,7 @@ const DuplicatesList = ({
   const { latestHttpCode: httpCodeEntry } = useSelector(state => state.entry);
 
   const { latestHttpCode: httpCodeDocument } = useSelector(
-    state => state.document
+    state => state.createDocument
   );
 
   const makeTranslation = id =>
@@ -75,9 +75,7 @@ const DuplicatesList = ({
 
   useEffect(() => {
     dispatch(resetEntranceState());
-    dispatch({
-      type: RESET_API_MESSAGES
-    });
+    dispatch(resetApiMessages());
   }, [httpCodeEntry, httpCodeDocument]);
 
   useEffect(() => {
