+++ packages/web-app/src/pages/Contributions/index.jsx
@@ -5,8 +5,8 @@ import { useIntl } from 'react-intl';
 import { reject, isNil, propOr, pathOr } from 'ramda';
 import { isMobileOnly } from 'react-device-detect';
 
-import { getUsersDocuments } from '../../actions/Documents';
-import { resetApiMessages } from '../../actions/Document';
+import { getUsersDocuments } from '../../actions/Document/GetDocuments';
+import { resetApiMessages } from '../../actions/Document/ResetApiMessages';
 import Layout from '../../components/common/Layouts/Fixed/FixedContent';
 import StandardDialog from '../../components/common/StandardDialog';
 import { useDebounce, useUserProperties } from '../../hooks';
